backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
gspread==5.12.3
oauth2client==4.1.3
Pillow==10.1.0
pydantic==2.5.2
pydantic-settings==2.1.0
aiofiles==23.2.1
groq==0.12.0
pymongo==4.6.1
openpyxl==3.1.2
cloudinary==1.36.0
